move class inside utils
