+++ antarest/study/storage/output_service.py
@@ -33,8 +33,8 @@ from antarest.core.utils.archives import ArchiveFormat
 from antarest.core.utils.fastapi_sqlalchemy import db
 from antarest.core.utils.utils import StopWatch
 from antarest.login.utils import get_user_id
-from antarest.study.business.output.aggregator_management import (
-    AggregatorManager,
+from antarest.study.business.output.aggregator_management import AggregatorManager
+from antarest.study.business.output.utils import (
     MCAllAreasQueryFile,
     MCAllLinksQueryFile,
     MCIndAreasQueryFile,
